Don't configure zope.component in every file.
This especially matters if the files will be loaded piecewise, e.g.,
in sublayers, like we do here.

Configuring zope.component is not idempotent (in general, configuring
any package that declares a subscriber is not idempotent). If we do
configure it again, we wind up with the same subscriber function
running twice for the same event. This wreaks havoc with things that
depend on those events, like setting up site libraries.



git-svn-id: https://repos.nextthought.com/svn/nti-svn/nti.app.client_preferences/trunk@98080 3dfb626b-079f-4ab1-a9f6-4654e5377508